requirements.txt
Django == 3.1.5

drf-yasg == 1.20.0

psycopg2 == 2.8.6

djangorestframework-simplejwt == 4.6.0

djangorestframework == 3.12.2